requirements.txt
altgraph==0.17.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
cycler==0.12.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
idna==3.10
jinja2==3.1.6
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pip==25.1.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.7
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.4
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
torch
torchvision
tqdm==4.67.1
typing-extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.128
ultralytics-thop==2.0.14
urllib3==2.5.0

